startgpts/startgpt/startgpt/core/pyproject.toml
[tool.poetry]
name = "agpt"
version = "1.0.0"
authors = ["Significant Gravitas <support@agpt.co>"]
maintainers = ["Reinier van der Leer <reinier.vanderleer@agpt.co>"]
description = "An open-source attempt at an autonomous generalist agent"
readme = "README.md"
repository = "https://github.com/KhulnaSoft/StartGPT/tree/master/startgpts/agpt"
# documentation = "https://docs.agpt.co/startgpts/agpt"  # TODO
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
packages = [{ include = "startgpt/core", from = "../.." }]

[tool.poetry.scripts]
cli = "startgpt.core.runner.cli_app.cli:startgpt"
cli-web = "startgpt.core.runner.cli_web_app.cli:startgpt"

[tool.poetry.dependencies]
python = "^3.10"
agent-protocol = "^0.3.0"
click = "^8.1.7"
colorama = "^0.4.6"
distro = "^1.8.0"
inflection = "^0.5.1"
jsonschema = "^4.19.1"
openai = "^0.28.0"
pydantic = "^1.10.12"
pyyaml = "^6.0.0"
tiktoken = "^0.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
packages = ["startgpt"]
extend-exclude = '.+/(dist|.venv|venv|build)/.+'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER"
]
skip = '''
    .tox
    __pycache__
    *.pyc
    .env
    venv*/*
    .venv/*
    reports/*
    dist/*

'''

[tool.pytest.ini_options]
markers = [
    "requires_openai_api_key",
    "requires_huggingface_api_key"
]
